Move StyledElements to seperate folder in UI
