+++ src/components/Pool/PairView/UserLiquidity/UserLiquidity.tsx
@@ -8,7 +8,7 @@ import styled from 'styled-components'
 import { useTotalSupply } from '../../../../data/TotalSupply'
 import { useActiveWeb3React } from '../../../../hooks'
 import { useTokenBalance } from '../../../../state/wallet/hooks'
-import { BlurBox } from '../../../../ui/BlurBox'
+import { BlurBox } from '../../../../ui/StyledElements/BlurBox'
 import { getAccountAnalyticsLink } from '../../../../utils'
 import { currencyId } from '../../../../utils/currencyId'
 import { unwrappedToken } from '../../../../utils/wrappedCurrency'
