+++ src/components/Pool/PairsList/LoadingList/LoadingList.tsx
@@ -3,9 +3,9 @@ import { useTranslation } from 'react-i18next'
 import { Flex } from 'rebass'
 
 import { useResponsiveItemsPerPage } from '../../../../hooks/useResponsiveItemsPerPage'
-import { ListLayout } from '../../../../ui/ListLayout'
-import { TableHeader } from '../../../../ui/TableHeader'
-import { TableHeaderText } from '../../../../ui/TableHeaderText'
+import { ListLayout } from '../../../../ui/StyledElements/ListLayout'
+import { TableHeader } from '../../../../ui/StyledElements/TableHeader'
+import { TableHeaderText } from '../../../../ui/StyledElements/TableHeaderText'
 import { LoadingRow } from './LoadingRow'
 
 export function LoadingList() {
