+++ src/components/Pool/PairsList/PairsList.tsx
@@ -13,10 +13,10 @@ import { useIsMobileByMedia } from '../../../hooks/useIsMobileByMedia'
 import { useNativeCurrencyUSDPrice } from '../../../hooks/useNativeCurrencyUSDPrice'
 import { usePage } from '../../../hooks/usePage'
 import { useResponsiveItemsPerPage } from '../../../hooks/useResponsiveItemsPerPage'
-import { BlurBox } from '../../../ui/BlurBox'
-import { ListLayout } from '../../../ui/ListLayout'
-import { TableHeader } from '../../../ui/TableHeader'
-import { TableHeaderText } from '../../../ui/TableHeaderText'
+import { BlurBox } from '../../../ui/StyledElements/BlurBox'
+import { ListLayout } from '../../../ui/StyledElements/ListLayout'
+import { TableHeader } from '../../../ui/StyledElements/TableHeader'
+import { TableHeaderText } from '../../../ui/StyledElements/TableHeaderText'
 import { getStakedAmountUSD } from '../../../utils/liquidityMining'
 import { ButtonPrimary } from '../../Button'
 import { Pagination } from '../../Pagination'
