+++ src/pages/GovPages/index.tsx
@@ -12,7 +12,7 @@ import { SearchInputWithIcon } from '../../components/SearchInputWithIcon'
 import { useNativeCurrency } from '../../hooks/useNativeCurrency'
 import { useRouter } from '../../hooks/useRouter'
 import { TYPE } from '../../theme'
-import { PageWrapper } from '../../ui/PageWrapper'
+import { PageWrapper } from '../../ui/StyledElements/PageWrapper'
 import { MainPage, PairPage } from './constant'
 import Container from './Container'
