+++ src/ui/StyledElements/BlurBox.tsx
@@ -1,7 +1,7 @@
 import { Box } from 'rebass'
 import styled from 'styled-components'
 
-import { gradients } from '../utils/theme'
+import { gradients } from '../../utils/theme'
 
 export const BlurBox = styled(Box)`
   border-radius: 12px;
